+++ ms-ne-emision-poliza-tps-pga/src/main/java/pe/com/pacifico/kuntur/expose/request/response/ProductoResponse.java
@@ -76,6 +76,6 @@ public class ProductoResponse implements Serializable {
   @NotBlank(message = "Input must not be empty")
   @ApiModelProperty(example = "19.90")
   @Pattern(regexp = "[0-9,]+[^.]", message = "Input should be only numbers")
-  private double precio;
+  private Long precio;
 
 }
