+++ ms-ne-emision-poliza-tps-pga/src/main/java/pe/com/pacifico/kuntur/model/Venta.java
@@ -37,7 +37,7 @@ public class Venta implements Serializable {
   private LocalDateTime fecVenta;
 
   @Column(name = "montototal")
-  private double montoTotal;
+  private Long montoTotal;
 
 
 }
